クレジットなどの差し替えした

+++ SummerProject/Banana.cpp
@@ -558,8 +558,8 @@ void Banana::Draw(Game& g)
 	auto	sx = _x - _w / 2;
 	auto	sy = _y - _h;
 
-	DrawGraph(sx, sy, _grHandle, TRUE);
+	DrawGraph(static_cast<int>(sx), static_cast<int>(sy), _grHandle, TRUE);
 	SetDrawBlendMode(DX_BLENDMODE_ALPHA, 128);
-	DrawBox(_x + _hit_x, _y + _hit_y, _x + _hit_x + _hit_w, _y + _hit_y + _hit_h, GetColor(255, 0, 0), TRUE);	// ”¼“§–¾‚ÌÔ‚Å“–‚½‚è”»’è•`‰æ
+	DrawBox(static_cast<int>(_x) + _hit_x, static_cast<int>(_y) + _hit_y, static_cast<int>(_x) + _hit_x + _hit_w, static_cast<int>(_y) + _hit_y + _hit_h, GetColor(255, 0, 0), TRUE);	// ”¼“§–¾‚ÌÔ‚Å“–‚½‚è”»’è•`‰æ
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);		// •s“§–¾•`‰æw’è
 }
