+++ SummerProject/Star.cpp
@@ -42,7 +42,7 @@ void Star::Init()
 	/*_x += _w / 2;
 	_y += _h;*/
 
-	while (_modeGame._newMapChips->IsHit(*this, _x, _y) != 0)
+	while (_modeGame._newMapChips->IsHit(*this, static_cast<int>(_x), static_cast<int>(_y)) != 0)
 	{
 		_x = rand() % _modeGame._newMapChips->MAPSIZE_W;
 		_y = rand() % _modeGame._newMapChips->MAPSIZE_H;
@@ -109,12 +109,12 @@ void Star::Process(Game& g)
 
 void Star::Draw(Game& g)
 {
-	auto	sx = _x - _w / 2;
-	auto	sy = _y - _h;
+	auto	sx = static_cast<int>(_x) - _w / 2;
+	auto	sy = static_cast<int>(_y) - _h;
 
 	DrawGraph(sx, sy, _grHandle, TRUE);
 	SetDrawBlendMode(DX_BLENDMODE_ALPHA, 128);
-	DrawBox(_x + _hit_x, _y + _hit_y, _x + _hit_x + _hit_w, _y + _hit_y + _hit_h, GetColor(255, 0, 0), TRUE);	// ”¼“§–¾‚ÌÔ‚Å“–‚½‚è”»’è•`‰æ
+	DrawBox(static_cast<int>(_x) + _hit_x, static_cast<int>(_y) + _hit_y, static_cast<int>(_x) + _hit_x + _hit_w, static_cast<int>(_y) + _hit_y + _hit_h, GetColor(255, 0, 0), TRUE);	// ”¼“§–¾‚ÌÔ‚Å“–‚½‚è”»’è•`‰æ
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);		// •s“§–¾•`‰æw’è
 }
